Feature/comment like

See merge request lbaw/lbaw2324/lbaw23141!37

+++ app/Http/Controllers/CommentController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 use App\Models\Comment;
+use App\Models\CommentLike;
 use App\Models\Post;
 
 use Illuminate\Auth\Access\AuthorizationException;
@@ -45,21 +46,38 @@ class CommentController extends Controller
       {  
         $comment = Comment::findOrFail($id);
         
-        Log::info($comment->comment_id);
+        Log::info($comment->author_id);
 
         $this->authorize('delete', $comment);
 
-        if ($comment) {
-            // Delete replies to the comment first
-            Comment::where('previous', $id)->delete();
-            // Delete the comment itself
-            $comment->delete();
-        }
-
+        $comment->delete();
+          
         return redirect()->back()->with('success', 'Comment deleted successfully!');
       }
       catch(AuthorizationException $e){
         return redirect()->back()->withErrors(['message' => 'You are not authorized to delete this comment']);
       }
     }
+
+    public function like($id) {
+      $authUser = Auth::user();
+      $comment = Comment::find($id);
+
+      $commentLike = new CommentLike([
+              'user_id' => $authUser->user_id,
+              'comment_id' => $comment->comment_id
+          ]);
+
+          $commentLike->save();
+      
+      return redirect()->back();
+  }
+
+  public function dislike($id) {
+    $authUser = Auth::user();
+    $comment = Comment::find($id);
+    $commentlike = CommentLike::where('user_id', $authUser->user_id)->where('comment_id', $comment->comment_id)->delete();
+
+    return redirect()->back();
+}
   }
